Refactor parallel functions
