+++ clowder/clowder/cli/herd_controller.py
@@ -7,7 +7,10 @@
 
 from cement.ext.ext_argparse import ArgparseController, expose
 
-from clowder.cli.parallel_commands import herd_parallel
+from clowder.cli.parallel_commands import (
+    herd_project,
+    run_parallel_command
+)
 from clowder.clowder_controller import CLOWDER_CONTROLLER
 from clowder.clowder_repo import print_clowder_repo_status_fetch
 from clowder.util.connectivity import network_connection_required
@@ -19,6 +22,7 @@ from clowder.util.clowder_utils import (
     run_group_command,
     run_project_command,
     validate_groups,
+    validate_print_output,
     validate_projects
 )
 
@@ -125,3 +129,35 @@ def herd(clowder, group_names, **kwargs):
     for project in projects:
         run_project_command(project, skip, 'herd', branch=branch, tag=tag,
                             depth=depth, rebase=rebase, protocol=protocol)
+
+
+def herd_parallel(clowder, group_names, **kwargs):
+    """Clone projects or update latest from upstream in parallel
+
+    .. py:function:: herd_parallel(clowder, group_names, branch=None, tag=None, depth=0, rebase=False, project_names=None, skip=[], protocol=None)
+
+    :param ClowderController clowder: ClowderController instance
+    :param list[str] group_names: Group names to herd
+
+    Keyword Args:
+        branch (str): Branch to attempt to herd
+        tag (str): Tag to attempt to herd
+        depth (int): Git clone depth. 0 indicates full clone, otherwise must be a positive integer
+        protocol (str): Git protocol ('ssh' or 'https')
+        rebase (bool): Whether to use rebase instead of pulling latest changes
+        project_names (list[str]): Project names to herd
+        skip (list[str]): Project names to skip
+    """
+
+    project_names = kwargs.get('project_names', None)
+    skip = kwargs.get('skip', [])
+    branch = kwargs.get('branch', None)
+    tag = kwargs.get('tag', None)
+    depth = kwargs.get('depth', None)
+    rebase = kwargs.get('rebase', False)
+    protocol = kwargs.get('protocol', None)
+
+    print(' - Herd projects in parallel\n')
+    validate_print_output(clowder, group_names, project_names=project_names, skip=skip)
+    projects = filter_projects(clowder.groups, group_names=group_names, project_names=project_names)
+    run_parallel_command(herd_project, projects, skip, branch, tag, depth, rebase, protocol)
