+++ clowder/clowder/cli/reset_controller.py
@@ -7,7 +7,10 @@
 
 from cement.ext.ext_argparse import ArgparseController, expose
 
-from clowder.cli.parallel_commands import reset_parallel
+from clowder.cli.parallel_commands import (
+    reset_project,
+    run_parallel_command
+)
 from clowder.clowder_controller import CLOWDER_CONTROLLER
 from clowder.clowder_repo import print_clowder_repo_status_fetch
 from clowder.util.connectivity import network_connection_required
@@ -19,6 +22,7 @@ from clowder.util.clowder_utils import (
     run_group_command,
     run_project_command,
     validate_groups,
+    validate_print_output,
     validate_projects
 )
 
@@ -112,3 +116,31 @@ def reset(clowder, group_names, **kwargs):
     validate_projects(projects)
     for project in projects:
         run_project_command(project, skip, 'reset', timestamp=timestamp)
+
+
+def reset_parallel(clowder, group_names, **kwargs):
+    """Reset project branches to upstream or checkout tag/sha as detached HEAD in parallel
+
+    .. py:function:: reset_parallel(clowder, group_names, timestamp_project=None, project_names=None, skip=[])
+
+    :param ClowderController clowder: ClowderController instance
+    :param list[str] group_names: Group names to reset
+
+    Keyword Args:
+        timestamp_project (str): Reference project to checkout commit timestamps of other projects relative to
+        project_names (list[str]): Project names to reset
+        skip (list[str]): Project names to skip
+    """
+
+    project_names = kwargs.get('project_names', None)
+    skip = kwargs.get('skip', [])
+    timestamp_project = kwargs.get('timestamp_project', None)
+
+    print(' - Reset projects in parallel\n')
+    timestamp = None
+    if timestamp_project:
+        timestamp = clowder.get_timestamp(timestamp_project)
+
+    validate_print_output(clowder, group_names, project_names=project_names, skip=skip)
+    projects = filter_projects(clowder.groups, group_names=group_names, project_names=project_names)
+    run_parallel_command(reset_project, projects, skip, timestamp)
