+++ clowder/clowder/cli/sync_controller.py
@@ -10,14 +10,18 @@ from __future__ import print_function
 from cement.ext.ext_argparse import ArgparseController, expose
 from termcolor import cprint
 
-from clowder.cli.parallel_commands import sync_parallel
+from clowder.cli.parallel_commands import (
+    run_parallel_command,
+    sync_project
+)
 from clowder.clowder_controller import CLOWDER_CONTROLLER
 from clowder.clowder_repo import print_clowder_repo_status_fetch
 from clowder.util.connectivity import network_connection_required
 from clowder.util.decorators import valid_clowder_yaml_required
 from clowder.util.clowder_utils import (
     filter_projects,
-    options_help_message
+    options_help_message,
+    print_parallel_projects_output
 )
 
 
@@ -79,3 +83,17 @@ def sync(clowder, project_names, rebase=False, parallel=False):
 
     for project in projects:
         project.sync(rebase=rebase)
+
+
+def sync_parallel(projects, rebase=False):
+    """Sync projects in parallel
+
+    .. py:function:: sync(projects, rebase=False)
+
+    :param list[Project] projects: Projects to sync
+    :param Optional[bool] rebase: Whether to use rebase instead of pulling latest changes
+    """
+
+    print(' - Sync forks in parallel\n')
+    print_parallel_projects_output(projects, [])
+    run_parallel_command(sync_project, projects, [], rebase)
